- Change the Makefiles to build the OpenGL version by default. Use GFX=software for the software renderer. - Still more warning fixes, this time mostly in lib/ivis02. Removed lots of Direct3D and otherwise unused code there in the process.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/branches/crossplatform@197 4a71c877-e1ca-e34f-864e-861f7616d084

+++ src/formation.c
@@ -788,8 +788,8 @@ BOOL formationGetPos( FORMATION *psFormation, BASE_OBJECT *psObj,
 	else if (psFormation->free != -1)
 	{
 		// add the new object to the members
-		psFormation->asMembers[psFormation->free].psObj = psObj;
-		psFormation->free = psFormation->asMembers[psFormation->free].next;
+		psFormation->asMembers[(int)psFormation->free].psObj = psObj;
+		psFormation->free = psFormation->asMembers[(int)psFormation->free].next;
 		formationReorder(psFormation);
 
 		// find the object
